requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.2
psycopg2-binary==2.9.9
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.3
yfinance==0.2.36
pytz==2024.1
aiohttp==3.9.3
